% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22031R.R
\name{data}
\alias{data}
\title{A illustration dataset}
\description{
A dataset used to illustrate the performance of \code{R_shortest} and \code{Rcpp_shortest}.
}
\examples{
\dontrun{
data(data)
attach(data)
benchmark(R_shortest(data),Rcpp_shortest(data),
columns=c("test","replications","elapsed","relative"))
}
}
