% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_cpp}
\alias{gibbs_cpp}
\title{A binary normal Gibbs sampler using Rcpp}
\usage{
gibbs_cpp(N, mu1, mu2, sigma1, sigma2, rho)
}
\arguments{
\item{N}{the sample size}

\item{mu1}{the mean value of one of the sample}

\item{mu2}{the mean value of another sample}

\item{sigma1}{the variance of one of the sample}

\item{sigma2}{the variance of another sample}

\item{rho}{the sample correlation coefficient}
}
\value{
A binary normal random sample of size \code{N}
}
\description{
Generate a binary normal sample with Gibbs
}
\examples{
\dontrun{
gibbs_cpp(5000,0,0,1,1,0.9)
}
}
