% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_shortest}
\alias{Rcpp_shortest}
\title{Shortest path model between cities}
\usage{
Rcpp_shortest(M)
}
\arguments{
\item{M}{a direct connection situation matrix of size \code{n*2}}
}
\value{
a matrix of size \code{n*n}
}
\description{
Output the connectivity between each city
}
\examples{
\dontrun{
data(data)
attach(data)
R_shortest(data)
}
}
