% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22031R.R
\name{s0_2}
\alias{s0_2}
\title{The value of initial stationary distribution}
\usage{
s0_2(kappa, delta, gamma, small)
}
\arguments{
\item{kappa}{the parameter of the distribution}

\item{delta}{the parameter of the distribution}

\item{gamma}{the parameter of the distribution}

\item{small}{the truncated variable}
}
\value{
a random sample of size \code{1}
}
\description{
Generate a sample of initial stationary distribution
}
\examples{
\dontrun{
sigma0_2=replicate(1000,s0_2(0.5,2,4,0.00001))
hist(sigma0_2,freq=FALSE)
}
}
