% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22031R.R
\name{s2_2}
\alias{s2_2}
\title{Compound Poission process driving part of TS-OU}
\usage{
s2_2(lambda, kappa, delta, gamma, Delta, small)
}
\arguments{
\item{lambda}{the parameter of the distribution}

\item{kappa}{the parameter of the distribution}

\item{delta}{the parameter of the distribution}

\item{gamma}{the parameter of the distribution}

\item{Delta}{the sampling interval}

\item{small}{the truncated variable}
}
\value{
a random sample of size \code{1}
}
\description{
Generate a sample of compound Poission driving part
}
\examples{
\dontrun{
sigma2_2=s2_2(0.1,0.4,2,5,1,0.00001)
}
}
