\name{StatComp22031-package}
\alias{StatComp22031-package}
\docType{package}
\title{
This package provides several functions for simulating TS-OU process and benchmarking R and Rcpp.
}
\description{
R function for simulating TS-OU process and R and Rcpp versions of the shortest path problem (also called acquaintance problem) function are provided in this package (\code{s0_2} and \code{s1_2} and \code{s2_2} and \code{TSOU} and \code{R_shortest} versus \code{Rcpp_shortest})
}
\author{
Mo Yang
Maintainer: Mo Yang <2691681557@qq.com>
}
\keyword{ package }
