% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22031R.R
\name{TSOU}
\alias{TSOU}
\title{The simulated path of TS-OU process}
\usage{
TSOU(lambda, kappa, delta, gamma, num, Delta = 1, small = 1e-05)
}
\arguments{
\item{lambda}{the parameter of the distribution}

\item{kappa}{the parameter of the distribution}

\item{delta}{the parameter of the distribution}

\item{gamma}{the parameter of the distribution}

\item{num}{the number of simulated samples}

\item{Delta}{the sampling interval}

\item{small}{the truncated variable}
}
\value{
a random sample of size \code{num}
}
\description{
Generate a sample of TS-OU process
}
\examples{
\dontrun{
sigma_2=TSOU(0.1,0.4,2,5,1000)
plot(sigma_2,type="l")
}
}
